Fix animatediff_motion_lora_pan_left: use bfloat16 to avoid L1 OOM on TT hardware
The AnimateDiff UNet (~860M params based on SD v1.5) causes L1 Out-of-Memory
(Error code: 13) when loaded in float32. Switch to bfloat16 (matching
stable_diffusion_unet which is EXPECTED_PASSING) to halve memory footprint.

+++ animatediff_motion_lora_pan_left/pytorch/loader.py
@@ -77,7 +77,7 @@ class ModelLoader(ForgeModel):
         Returns:
             UNetMotionModel (torch.nn.Module) with LoRA weights merged.
         """
-        dtype = dtype_override if dtype_override is not None else torch.float32
+        dtype = dtype_override if dtype_override is not None else torch.bfloat16
 
         adapter = MotionAdapter.from_pretrained(
             MOTION_ADAPTER,
@@ -102,7 +102,7 @@ class ModelLoader(ForgeModel):
         Returns:
             dict with sample, timestep, and encoder_hidden_states tensors.
         """
-        dtype = dtype_override if dtype_override is not None else torch.float32
+        dtype = dtype_override if dtype_override is not None else torch.bfloat16
 
         batch_size = 1
         num_frames = 32  # must be >= 32: TT hardware SDPA requires key seq_len >= 32
